+++ backend/internal/handler/company/delete_company.go
@@ -1,0 +1,41 @@
+package company
+
+import (
+	"net/http"
+	"strconv"
+
+	"github.com/gin-gonic/gin"
+	"github.com/takanoakira/ai-interview-practice/backend/internal/usecase/company"
+)
+
+// DeleteCompany は指定されたIDの企業情報を削除するハンドラーです
+func DeleteCompany(companyUseCase company.CompanyUseCase) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		// パスパラメータからIDを取得
+		idStr := c.Param("id")
+		id, err := strconv.Atoi(idStr)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{
+				"error": gin.H{
+					"code":    "INVALID_ID",
+					"message": "IDは整数である必要があります",
+				},
+			})
+			return
+		}
+
+		// ユースケースを呼び出し
+		if err := companyUseCase.DeleteCompany(c.Request.Context(), id); err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{
+				"error": gin.H{
+					"code":    "SERVER_ERROR",
+					"message": "サーバーエラーが発生しました",
+				},
+			})
+			return
+		}
+
+		// 成功レスポンスを返す
+		c.Status(http.StatusNoContent)
+	}
+}
